+++ R/heatplot.R
@@ -2,14 +2,16 @@
 ##' @exportMethod heatplot
 setMethod("heatplot", signature(x = "enrichResult"),
           function(x, showCategory = 30, foldChange = NULL) {
-              heatplot.enrichResult(x, showCategory, foldChange)
+              heatplot.enrichResult(x, showCategory, foldChange,
+                                    label_format = 30)
           })
 
 ##' @rdname heatplot
 ##' @exportMethod heatplot
 setMethod("heatplot", signature(x = "gseaResult"),
           function(x, showCategory = 30, foldChange = NULL) {
-              heatplot.enrichResult(x, showCategory, foldChange)
+              heatplot.enrichResult(x, showCategory, foldChange,
+                                    label_format = 30)
           })
 
 
@@ -20,8 +22,19 @@ setMethod("heatplot", signature(x = "gseaResult"),
 ##' @importFrom ggplot2 theme
 ##' @importFrom ggplot2 element_blank
 ##' @importFrom ggplot2 element_text
+##' @importFrom ggplot2 scale_y_discrete
+##' @param label_format a numeric value sets wrap length, alternatively a
+##' custom function to format axis labels.
+##' by default wraps names longer that 30 characters
 ##' @author Guangchuang Yu
-heatplot.enrichResult <- function(x, showCategory=30, foldChange=NULL) {
+heatplot.enrichResult <- function(x, showCategory=30, foldChange=NULL,
+                                  label_format = 30) {
+
+    label_func <- default_labeller(label_format)
+    if(is.function(label_format)) {
+        label_func <- label_format
+    }
+
     n <- update_n(x, showCategory)
     geneSets <- extract_geneSets(x, n)
 
@@ -40,6 +53,7 @@ heatplot.enrichResult <- function(x, showCategory=30, foldChange=NULL) {
         p <- ggplot(d, aes_(~Gene, ~categoryID)) + geom_tile(color = 'white')
     }
     p + xlab(NULL) + ylab(NULL) + theme_minimal() +
+        scale_y_discrete(labels = label_func) +
         theme(panel.grid.major = element_blank(),
               axis.text.x=element_text(angle = 60, hjust = 1))
 }